Also pass --path to commands that take stdin
Summary: One of the downsides to passing the buffer via stdin and not via a
pathname is that Flow doesn't know what the root is anymore, so it assumes that
your cwd is in the root. So if you edit `~/my/project/foo.js` with your cwd as
`~`, Flow won't find the `~/my/project/.flowconfig`

The easy fix is to tell Flow what file you're editing.

Test Plan: Edited a Flow repo from outside the Flow repo. Then ran

  :FlowType
  :FlowJumpToDef

previously these errored. Now they work.